+++ packages/cfd/src/Containers/mt5-mobile-redirect-option.scss
